chore: update namespaces after reorganization
