+++ Backend/src/Modules/Nodes/Nodes.Application/Entities/Nodes/Commands/CreateNode/CreateNodeHandler.cs
@@ -1,9 +1,8 @@
 using Nodes.Application.Data;
 
-namespace Nodes.Application.Nodes.Commands.CreateNode;
+namespace Nodes.Application.Entities.Nodes.Commands.CreateNode;
 
-public class CreateNodeHandler(INodesDbContext dbContext, IValidator<CreateNodeCommand> validator) :
-    ICommandHandler<CreateNodeCommand, CreateNodeResult>
+public class CreateNodeHandler(INodesDbContext dbContext) : ICommandHandler<CreateNodeCommand, CreateNodeResult>
 {
     public async Task<CreateNodeResult> Handle(CreateNodeCommand command, CancellationToken cancellationToken)
     {
