+++ Backend/src/Modules/Nodes/Nodes.Application/Entities/Nodes/Exceptions/NodeNotFoundException.cs
@@ -1,3 +1,3 @@
-namespace Nodes.Application.Exceptions;
+namespace Nodes.Application.Entities.Nodes.Exceptions;
 
 public class NodeNotFoundException(string id) : NotFoundException("Node", id);
