+++ Backend/src/Modules/Nodes/Nodes.Application/GlobalUsing.cs
@@ -3,5 +3,4 @@ global using Microsoft.EntityFrameworkCore;
 global using BuildingBlocks.Application.Cqrs;
 global using BuildingBlocks.Application.Exceptions;
 global using BuildingBlocks.Domain.ValueObjects.Ids;
-global using Nodes.Application.Dtos;
 global using Nodes.Domain.Models;
